持久化相关代码提交
git-svn-id: http://192.168.20.11/sequoiadb/trunk@24665 6121f079-4a18-41d3-8c30-8870a322f2c2

+++ SequoiaDB/engine/cls/clsFSDstSession.cpp
@@ -1566,7 +1566,8 @@ namespace engine
    {
       if ( CLS_FS_NOTIFY_TYPE_LOB == pMsg->type &&
            CLS_FS_EOF == pMsg->eof &&
-           pMsg->header.header.messageLength > sizeof( MsgClsFSNotifyRes ) )
+           (UINT32)(pMsg->header.header.messageLength) >
+           sizeof( MsgClsFSNotifyRes ) )
       {
          rtnRUInfo info ;
          const CHAR *pCLName = _fullNames[_current].c_str() ;
